migration(micronauts): export Module api to separate controller
